--- compiler/frontend/src/mir_optimize/base.rs
@@ -1,6 +1,0 @@
-use crate::mir::{Expression, VisibleExpressions};
-
-pub trait ExpressionOptimization {
-    fn name(&self) -> &'static str;
-    fn apply(&self, expression: &mut Expression, visible: &VisibleExpressions);
-}
