+++ compiler/frontend/src/mir_optimize/constant_folding.rs
@@ -34,11 +34,11 @@ use num_traits::ToPrimitive;
 use crate::{
     builtin_functions::BuiltinFunction,
     id::IdGenerator,
-    mir::{Body, Expression, Id, Mir, VisibleExpressions},
+    mir::{Body, Expression, Id, VisibleExpressions},
     rich_ir::ToRichIr,
 };
 
-pub fn apply(
+pub fn fold_constants(
     expression: &mut Expression,
     visible: &VisibleExpressions,
     id_generator: &mut IdGenerator<Id>,
@@ -70,46 +70,10 @@ pub fn apply(
     *expression = evaluated_call;
 }
 
-impl Mir {
-    pub fn fold_constants(&mut self) {
-        self.body
-            .visit_with_visible(&mut |_, expression, visible, _| {
-                let Expression::Call {
-                    function,
-                    arguments,
-                    responsible,
-                } = expression else { return; };
-                let Expression::Builtin(builtin) = visible.get(*function) else { return; };
-                let Some(result) = run_builtin(*builtin, arguments, *responsible, visible) else {
-                    return;
-                };
-                let evaluated_call = match result {
-                    Ok(return_value) => return_value,
-                    Err(panic_reason) => {
-                        let mut body = Body::default();
-                        let reason = body.push_with_new_id(
-                            &mut self.id_generator,
-                            Expression::Text(panic_reason),
-                        );
-                        body.push_with_new_id(
-                            &mut self.id_generator,
-                            Expression::Panic {
-                                reason,
-                                responsible: *responsible,
-                            },
-                        );
-                        Expression::Multiple(body)
-                    }
-                };
-                *expression = evaluated_call;
-            });
-    }
-}
-
-/// This function tries to run a builtin, requiring a minimal amount of
-/// static knowledge. For example, it can find out that the result of
-/// `builtinEquals $3 $3` is `True`, even if the value of `$3` is not known
-/// at compile-time.
+/// This function tries to run a builtin, requiring a minimal amount of static
+/// knowledge. For example, it can find out that the result of
+/// `builtinEquals $3 $3` is `True`, even if the value of `$3` is not known at
+/// compile-time.
 ///
 /// Returns `None` if the call couldn't be evaluated statically. Returns
 /// `Some(Ok(expression))` if the call successfully completed with a return
