+++ compiler/frontend/src/mir_optimize/reference_following.rs
@@ -16,11 +16,9 @@
 //!
 //! [constant folding]: super::constant_folding
 
-use rustc_hash::FxHashMap;
+use crate::mir::{Body, Expression, VisibleExpressions};
 
-use crate::mir::{Body, Expression, Mir, VisibleExpressions, VisitorResult};
-
-pub fn apply(expression: &mut Expression, visible: &VisibleExpressions) {
+pub fn follow_references(expression: &mut Expression, visible: &VisibleExpressions) {
     expression.replace_id_references(&mut |id| {
         if visible.contains(*id) && let Expression::Reference(referenced) = visible.get(*id) {
             *id = *referenced;
@@ -40,44 +38,5 @@ pub fn remove_redundant_return_references(body: &mut Body) {
         } else {
             break;
         }
-    }
-}
-
-impl Mir {
-    pub fn follow_references(&mut self) {
-        let mut replacements = FxHashMap::default();
-
-        self.body.visit(&mut |id, expression, _| {
-            if let Expression::Reference(reference) = &expression {
-                let replacement = *replacements.get(reference).unwrap_or(reference);
-                replacements.insert(id, replacement);
-            }
-            VisitorResult::Continue
-        });
-        self.body.visit(&mut |_, expression, _| {
-            expression.replace_id_references(&mut |id| {
-                if let Some(&replacement) = replacements.get(id) {
-                    *id = replacement;
-                }
-            });
-            VisitorResult::Continue
-        });
-    }
-
-    pub fn remove_redundant_return_references(&mut self) {
-        self.body.visit_bodies(&mut |body| {
-            loop {
-                let mut from_back = body.iter_mut().rev();
-                let (last_id, last_expression) = from_back.next().unwrap();
-                let Some((before_last_id, _)) = from_back.next() else { return; };
-
-                if let Expression::Reference(referenced) = last_expression && before_last_id == *referenced {
-                    drop(from_back);
-                    body.remove_all(|id, _| last_id == id);
-                } else {
-                    break;
-                }
-            }
-        });
     }
 }
